Another layout update
Details: Chamber will only bee a bit bigger than it currently is.
Current room containing red.. thing is a decoy (similar to fake rooms
installed in real pyramids.) The chamber is the real burial room, and
where the boss will be fought post-plantera. There is now a "door"
blocking progression through the pyramid until the player defeats the
boss.